Refactor deadlock logic
Now deadlock warning log will be reported twice, the reason of it is that we set a defer task in the next round to unlock the deadlock, but uv event loop uv_backend_timeout(loop) is 0 for the first loop, so it was unlocked then it continues to the next deadlock. I can not solve this now, and use forever timer to implement deadlock seems unnecessary, so we just use sleep to do that, it can be interrupted by set COROUTINE_G(deadlocked) for debugging.