[BC] New hand-written media-type parser (#23)
This is still handwritten code, and its performance is comparable to the original version before modification. It's slightly faster than the one generated by Ragel.
It conflicts with issue #22.

BC:
When the Content-Type is empty, the parser will no longer report an error if there are duplicates, but instead, continue parsing without enabling the multipart parser.
Most of the error codes related to multipart have been removed (although they can be easily added back compared to Ragel).
The parsing of boundary has become stricter and is now mostly in accordance with the relevant RFC.

Related: swow/swow#186.